pyproject.toml
[project]
name = "clashbot"
version = "0.2.0"
description = "A Discord bot for tracking Clash of Clans activity."
authors = [
    {name = "vrtx",email = "vortexshrimp@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py (>=2.5.2,<3.0.0)",
    "coc-py (>=3.9.1,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
